Add files via upload
There is supposed to be 10,000 results of a dieroll with 3 dice. They should all be in a list, and then in brievaQ1 have the average. In brievaQ2 you need to have the numbers print out like 08-2, 09-3, 10-1, which represents the number and the amount of times rolled. You also need the standard deviation of this. Then in brievaQ3, you have to code a line graph with the results of the die roll.